+++ mobile/library/objective-c/EnvoyConfiguration.h
@@ -20,9 +20,13 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, strong) NSString *dnsPreresolveHostnames;
 @property (nonatomic, assign) UInt32 dnsRefreshSeconds;
 @property (nonatomic, assign) BOOL enableDNSCache;
+@property (nonatomic, assign) UInt32 dnsCacheSaveIntervalSeconds;
 @property (nonatomic, assign) BOOL enableHappyEyeballs;
-@property (nonatomic, assign) BOOL enableGzip;
-@property (nonatomic, assign) BOOL enableBrotli;
+@property (nonatomic, assign) BOOL enableHttp3;
+@property (nonatomic, assign) BOOL enableGzipDecompression;
+@property (nonatomic, assign) BOOL enableGzipCompression;
+@property (nonatomic, assign) BOOL enableBrotliDecompression;
+@property (nonatomic, assign) BOOL enableBrotliCompression;
 @property (nonatomic, assign) BOOL enableInterfaceBinding;
 @property (nonatomic, assign) BOOL enableDrainPostDnsRefresh;
 @property (nonatomic, assign) BOOL enforceTrustChainVerification;
@@ -58,9 +62,13 @@ NS_ASSUME_NONNULL_BEGIN
                              dnsMinRefreshSeconds:(UInt32)dnsMinRefreshSeconds
                            dnsPreresolveHostnames:(NSString *)dnsPreresolveHostnames
                                    enableDNSCache:(BOOL)enableDNSCache
+                      dnsCacheSaveIntervalSeconds:(UInt32)dnsCacheSaveIntervalSeconds
                               enableHappyEyeballs:(BOOL)enableHappyEyeballs
-                                       enableGzip:(BOOL)enableGzip
-                                     enableBrotli:(BOOL)enableBrotli
+                                      enableHttp3:(BOOL)enableHttp3
+                          enableGzipDecompression:(BOOL)enableGzipDecompression
+                            enableGzipCompression:(BOOL)enableGzipCompression
+                        enableBrotliDecompression:(BOOL)enableBrotliDecompression
+                          enableBrotliCompression:(BOOL)enableBrotliCompression
                            enableInterfaceBinding:(BOOL)enableInterfaceBinding
                         enableDrainPostDnsRefresh:(BOOL)enableDrainPostDnsRefresh
                     enforceTrustChainVerification:(BOOL)enforceTrustChainVerification