+++ mobile/library/objective-c/EnvoyConfiguration.m
@@ -14,9 +14,13 @@
                              dnsMinRefreshSeconds:(UInt32)dnsMinRefreshSeconds
                            dnsPreresolveHostnames:(NSString *)dnsPreresolveHostnames
                                    enableDNSCache:(BOOL)enableDNSCache
+                      dnsCacheSaveIntervalSeconds:(UInt32)dnsCacheSaveIntervalSeconds
                               enableHappyEyeballs:(BOOL)enableHappyEyeballs
-                                       enableGzip:(BOOL)enableGzip
-                                     enableBrotli:(BOOL)enableBrotli
+                                      enableHttp3:(BOOL)enableHttp3
+                          enableGzipDecompression:(BOOL)enableGzipDecompression
+                            enableGzipCompression:(BOOL)enableGzipCompression
+                        enableBrotliDecompression:(BOOL)enableBrotliDecompression
+                          enableBrotliCompression:(BOOL)enableBrotliCompression
                            enableInterfaceBinding:(BOOL)enableInterfaceBinding
                         enableDrainPostDnsRefresh:(BOOL)enableDrainPostDnsRefresh
                     enforceTrustChainVerification:(BOOL)enforceTrustChainVerification
@@ -60,9 +64,13 @@
   self.dnsMinRefreshSeconds = dnsMinRefreshSeconds;
   self.dnsPreresolveHostnames = dnsPreresolveHostnames;
   self.enableDNSCache = enableDNSCache;
+  self.dnsCacheSaveIntervalSeconds = dnsCacheSaveIntervalSeconds;
   self.enableHappyEyeballs = enableHappyEyeballs;
-  self.enableGzip = enableGzip;
-  self.enableBrotli = enableBrotli;
+  self.enableHttp3 = enableHttp3;
+  self.enableGzipDecompression = enableGzipDecompression;
+  self.enableGzipCompression = enableGzipCompression;
+  self.enableBrotliDecompression = enableBrotliDecompression;
+  self.enableBrotliCompression = enableBrotliCompression;
   self.enableInterfaceBinding = enableInterfaceBinding;
   self.enableDrainPostDnsRefresh = enableDrainPostDnsRefresh;
   self.enforceTrustChainVerification = enforceTrustChainVerification;
@@ -112,14 +120,42 @@
     [customFilters appendString:filterConfig];
   }
 
-  if (self.enableGzip) {
-    NSString *gzipFilterInsert = [[NSString alloc] initWithUTF8String:gzip_config_insert];
-    [customFilters appendString:gzipFilterInsert];
+  if (self.enableGzipDecompression) {
+    NSString *insert = [[NSString alloc] initWithUTF8String:gzip_decompressor_config_insert];
+    [customFilters appendString:insert];
   }
 
-  if (self.enableBrotli) {
-    NSString *brotliFilterInsert = [[NSString alloc] initWithUTF8String:brotli_config_insert];
-    [customFilters appendString:brotliFilterInsert];
+  if (self.enableGzipCompression) {
+#if ENVOY_MOBILE_REQUEST_COMPRESSION
+    NSString *insert = [[NSString alloc] initWithUTF8String:gzip_compressor_config_insert];
+    [customFilters appendString:insert];
+#else
+    NSLog(@"[Envoy] error: request compression functionality was not compiled in this build of "
+          @"Envoy Mobile");
+    return nil;
+#endif
+  }
+
+  if (self.enableBrotliDecompression) {
+    NSString *insert = [[NSString alloc] initWithUTF8String:brotli_decompressor_config_insert];
+    [customFilters appendString:insert];
+  }
+
+  if (self.enableBrotliCompression) {
+#if ENVOY_MOBILE_REQUEST_COMPRESSION
+    NSString *insert = [[NSString alloc] initWithUTF8String:brotli_compressor_config_insert];
+    [customFilters appendString:insert];
+#else
+    NSLog(@"[Envoy] error: request compression functionality was not compiled in this build of "
+          @"Envoy Mobile");
+    return nil;
+#endif
+  }
+
+  if (self.enableHttp3) {
+    NSString *http3Insert =
+        [[NSString alloc] initWithUTF8String:alternate_protocols_cache_filter_insert];
+    [customFilters appendString:http3Insert];
   }
 
   BOOL hasDirectResponses = self.directResponses.length > 0;
@@ -159,8 +195,6 @@
   [definitions appendFormat:@"- &dns_preresolve_hostnames %@\n", self.dnsPreresolveHostnames];
   [definitions appendFormat:@"- &dns_lookup_family %@\n",
                             self.enableHappyEyeballs ? @"ALL" : @"V4_PREFERRED"];
-  [definitions appendFormat:@"- &dns_multiple_addresses %@\n",
-                            self.enableHappyEyeballs ? @"true" : @"false"];
   [definitions appendFormat:@"- &dns_refresh_rate %lus\n", (unsigned long)self.dnsRefreshSeconds];
   [definitions appendFormat:@"- &enable_drain_post_dns_refresh %@\n",
                             self.enableDrainPostDnsRefresh ? @"true" : @"false"];
@@ -194,6 +228,8 @@
   [definitions appendFormat:@"%@\n", cert_validator_template];
 
   if (self.enableDNSCache) {
+    [definitions appendFormat:@"- &persistent_dns_cache_save_interval %lu\n",
+                              (unsigned long)self.dnsCacheSaveIntervalSeconds];
     NSString *persistent_dns_cache_config = @(persistent_dns_cache_config_insert);
     [definitions appendFormat:@"- &persistent_dns_cache_config %@\n", persistent_dns_cache_config];
   }